requirements.txt
azure-storage-blob==12.17.0
azure-search-documents==11.4.0
openai==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
tiktoken==0.5.1
langchain==0.1.0
pydantic==2.5.0
numpy==1.26.0
tqdm==4.66.1
colorama==0.4.6
click==8.1.7
tenacity==8.2.3